+++ lib/capistrano/puma.rb
@@ -1,1 +1,5 @@
-load File.expand_path('../tasks/puma.cap', __FILE__)
+require 'capistrano/puma/utility'
+
+include Capistrano::Puma::Utility
+
+load File.expand_path('../puma/tasks/puma.cap', __FILE__)