+++ lib/capistrano/puma/workers.rb
@@ -1,1 +1,3 @@
-load File.expand_path('../../tasks/workers.cap', __FILE__)
+require 'capistrano/puma'
+
+load File.expand_path('../tasks/workers.cap', __FILE__)